Modified Login system and DCS register

+++ src/dbLogin.sh
@@ -90,14 +90,14 @@ if [ ! -f ${cfg} ]; then
     fi
 
     echo " "
-    echo "You can set the identification key (e.x. nickname, SW version you use ...)"
-    echo "Do you want to set the identification key? [y/n]"
-    echo "    y => continue to set the identification key"
-    echo "    n => set the identification key 'default'"
+    echo "You can set the identification keyword if you want. (e.x. nickname, SW version you use ...)"
+    echo "Do you want to set the identification keyword? [y/n]"
+    echo "    y => continue to set the identification keyword"
+    echo "    n => set the identification keyword 'default'"
     read -p "> " -a answer
     while [ ${#answer[@]} == 0 ]; 
     do
-        echo "Enter 'y' to set the identification key, 'n' not to set, or 'exit' to abort ... "
+        echo "Enter 'y' to set the identification keyword, 'n' not to set, or 'exit' to abort ... "
         read -p "> " -a answer
     done
     if [ ${answer[0]} == "exit" ]; then
@@ -109,11 +109,11 @@ if [ ! -f ${cfg} ]; then
         identity="default"
     else
         echo " "
-        echo "Enter your identification key or 'cancel' ... "
+        echo "Enter your identification keyword or 'cancel' ... "
         read -p "> " -a answer
         while [ ${#answer[@]} == 0 ]; 
         do
-            echo "Enter your identification key or 'cancel' ... "
+            echo "Enter your identification keyword or 'cancel' ... "
             read -p "> " -a answer
         done
         if [ ${answer[0]} == "cancel" ]; then
@@ -149,12 +149,16 @@ done
 dbcfg=${HOME}/.yarr/database.json
 
 if [ ${answer} == "y" ]; then
-    echo "{" > ${cfg}
-    echo "    \"userName\": \"${name}\"," >> ${cfg}
-    echo "    \"institution\": \"${institution}\"," >> ${cfg}
-    echo "    \"userIdentity\": \"${identity}\"," >> ${cfg}
-    echo "    \"dbCfg\": \"${dbcfg}\"" >> ${cfg}
-    echo "}" >> ${cfg}
+    if [ ! -f ${cfg} ]; then
+        echo "{" > ${cfg}
+        echo "    \"userName\": \"${name}\"," >> ${cfg}
+        echo "    \"institution\": \"${institution}\"," >> ${cfg}
+        echo "    \"userIdentity\": \"${identity}\"," >> ${cfg}
+        echo "    \"dbCfg\": \"${dbcfg}\"" >> ${cfg}
+        echo "}" >> ${cfg}
+        echo " "
+        echo "Create User Config file: ${cfg}"
+    fi
     echo " "
     if "${DEBUG}"; then
         echo "export DBUSER=\"${account}\""
@@ -166,8 +170,6 @@ if [ ${answer} == "y" ]; then
         echo " "
     fi
     ./bin/dbAccessor -U ${account}
-    echo "Create User Config file: ${cfg}"
-    echo " "
 else
     echo "Exit ..."
     echo " "
@@ -175,40 +177,40 @@ else
     return 0
 fi
 
-if [ ! -f ${dbcfg} ]; then
-    echo "{" > ${dbcfg}
-    echo "    \"stage\": [" >> ${dbcfg}
-    echo "        \"Bare Module\"," >> ${dbcfg}
-    echo "        \"Wire Bonded\"," >> ${dbcfg}
-    echo "        \"Potted\"," >> ${dbcfg}
-    echo "        \"Final Electrical\"," >> ${dbcfg}
-    echo "        \"Complete\"," >> ${dbcfg}
-    echo "        \"Loaded\"," >> ${dbcfg}
-    echo "        \"Parylene\"," >> ${dbcfg}
-    echo "        \"Initial Electrical\"," >> ${dbcfg}
-    echo "        \"Thermal Cycling\"," >> ${dbcfg}
-    echo "        \"Flex + Bare Module Attachment\"," >> ${dbcfg}
-    echo "        \"Testing\"" >> ${dbcfg}
-    echo "    ]," >> ${dbcfg}
-    echo "    \"environment\": [" >> ${dbcfg}
-    echo "        \"lv\"," >> ${dbcfg}
-    echo "        \"hv\"," >> ${dbcfg}
-    echo "        \"lv_current\"," >> ${dbcfg}
-    echo "        \"hv_current\"," >> ${dbcfg}
-    echo "        \"temperature\"" >> ${dbcfg}
-    echo "    ]," >> ${dbcfg}
-    echo "    \"component\": [" >> ${dbcfg}
-    echo "        \"Front-end Chip\"," >> ${dbcfg}
-    echo "        \"Front-end Chips Wafer\"," >> ${dbcfg}
-    echo "        \"Hybrid\"," >> ${dbcfg}
-    echo "        \"Module\"," >> ${dbcfg}
-    echo "        \"Sensor Tile\"," >> ${dbcfg}
-    echo "        \"Sensor Wafer\"" >> ${dbcfg}
-    echo "    ]" >> ${dbcfg}
-    echo "}" >> ${dbcfg}
-    echo "Create DB Config file: ${dbcfg}"
-    echo " "
-fi
+echo "{" > ${dbcfg}
+echo "    \"stage\": [" >> ${dbcfg}
+echo "        \"Bare Module\"," >> ${dbcfg}
+echo "        \"Wire Bonded\"," >> ${dbcfg}
+echo "        \"Potted\"," >> ${dbcfg}
+echo "        \"Final Electrical\"," >> ${dbcfg}
+echo "        \"Complete\"," >> ${dbcfg}
+echo "        \"Loaded\"," >> ${dbcfg}
+echo "        \"Parylene\"," >> ${dbcfg}
+echo "        \"Initial Electrical\"," >> ${dbcfg}
+echo "        \"Thermal Cycling\"," >> ${dbcfg}
+echo "        \"Flex + Bare Module Attachment\"," >> ${dbcfg}
+echo "        \"Testing\"" >> ${dbcfg}
+echo "    ]," >> ${dbcfg}
+echo "    \"environment\": [" >> ${dbcfg}
+echo "        \"vddd_voltage\"," >> ${dbcfg}
+echo "        \"vddd_current\"," >> ${dbcfg}
+echo "        \"vdda_voltage\"," >> ${dbcfg}
+echo "        \"vdda_current\"," >> ${dbcfg}
+echo "        \"hv_voltage\"," >> ${dbcfg}
+echo "        \"hv_current\"," >> ${dbcfg}
+echo "        \"temperature\"" >> ${dbcfg}
+echo "    ]," >> ${dbcfg}
+echo "    \"component\": [" >> ${dbcfg}
+echo "        \"Front-end Chip\"," >> ${dbcfg}
+echo "        \"Front-end Chips Wafer\"," >> ${dbcfg}
+echo "        \"Hybrid\"," >> ${dbcfg}
+echo "        \"Module\"," >> ${dbcfg}
+echo "        \"Sensor Tile\"," >> ${dbcfg}
+echo "        \"Sensor Wafer\"" >> ${dbcfg}
+echo "    ]" >> ${dbcfg}
+echo "}" >> ${dbcfg}
+echo "Create DB Config file: ${dbcfg}"
+echo " "
 
 declare -a nic=()  
 num=0
